apps/siwa-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6

pydantic[email]==2.9.2
pydantic-settings==2.5.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

python-multipart==0.0.9

passlib[bcrypt]==1.7.4
bcrypt==3.2.2

pillow==10.4.0
numpy==1.26.4
pydicom==2.4.4

torch==2.4.1
requests==2.32.3
scikit-learn==1.4.0
evaluate==0.4.6
bert-score==0.3.13
sacrebleu==2.5.1
rouge-score==0.1.2
nltk==3.9.2

sqlalchemy
